Cleanup code, implement mocks and make build pass

+++ service/history/historyEngine.go
@@ -2338,6 +2338,6 @@ func getWorkflowAlreadyStartedError(errMsg string, createRequestID string, workf
 	}
 }
 
-func (e *historyEngineImpl) GetReplicationTasks(ctx ctx.Context, taskID int) ([]*replicator.ReplicationTask, error) {
+func (e *historyEngineImpl) GetReplicationTasks(ctx ctx.Context, taskID int64) ([]*replicator.ReplicationTask, error) {
 	return e.replicatorProcessor.getTasks(taskID)
 }
